(core) audio-id reuse fixes
FossilOrigin-Name: 84a2429c3f0b4033900e73586efd1b23d87408b2734bfe15912d33da53268002

+++ engine/arcan_audioint.h
@@ -74,7 +74,7 @@ typedef struct arcan_aobj {
 } arcan_aobj;
 
 /* just a wrapper around alBufferData that takes monitors into account */
-void arcan_audio_buffer(arcan_aobj*, unsigned, void*,
+void arcan_audio_buffer(arcan_aobj*, ssize_t, void*,
 	size_t, unsigned, unsigned, void*);
 
 #endif
